Extract Odoo CLI commands
